@@ -7,20 +7,20 @@ import 'package:permission_handler/permission_handler.dart' as handler;
 import 'package:location_permissions/location_permissions.dart' as location_perm;
 
 class LocationManager {
-  LatLng currentLocation;
+  late LatLng currentLocation;
   Location location = Location();
-  Function(LatLng) onLocationValue;
-  VoidCallback onRejectPermission;
-  bool isLocationTrack;
+  late Function(LatLng) onLocationValue;
+  late VoidCallback onRejectPermission;
+  late bool isLocationTrack;
 
-  StreamSubscription<LocationData> locationManager;
+  late StreamSubscription<LocationData> locationManager;
 
-  Future<void> getLocation({Function(LatLng) onLocationValue, VoidCallback onRejectPermission, bool isLocationTrack = false, int interval = 1000}) async {
+  Future<void> getLocation({required Function(LatLng) onLocationValue, required VoidCallback onRejectPermission, bool isLocationTrack = false, int interval = 1000}) async {
     if (await getPermissionStatus()) {
       var serviceStatus = await location_perm.LocationPermissions().checkServiceStatus();
 
       if (serviceStatus != location_perm.ServiceStatus.enabled) {
-        await location_perm.LocationPermissions().serviceStatus.listen((event) {
+        location_perm.LocationPermissions().serviceStatus.listen((event) {
           if (event.index == 2) {
             getLocation(onLocationValue: onLocationValue, isLocationTrack: isLocationTrack, onRejectPermission: onRejectPermission);
           } else {
@@ -39,20 +39,20 @@ class LocationManager {
     }
   }
 
-  Future<void> getCurrentLocation({Function(LatLng) onLocationValue}) async {
+  Future<void> getCurrentLocation({required Function(LatLng) onLocationValue}) async {
     var _locationResult = await location.getLocation();
-    onLocationValue(LatLng(_locationResult.latitude, _locationResult.longitude));
+    onLocationValue(LatLng(_locationResult.latitude!, _locationResult.longitude!));
   }
 
-  Future<void> listenLocation({Function(LatLng) onLocationValue, int interval}) async {
+  Future<void> listenLocation({required Function(LatLng) onLocationValue, int interval = 3000}) async {
     await location.changeSettings(interval: interval);
-    locationManager = await location.onLocationChanged.listen((LocationData currentLocation) {
+    locationManager = location.onLocationChanged.listen((LocationData currentLocation) {
       print('listening...');
     });
 
     locationManager.onData((data) {
       try {
-        onLocationValue(LatLng(data.latitude, data.longitude));
+        onLocationValue(LatLng(data.latitude!, data.longitude!));
       } catch (e) {
         print("listenLocation catch");
       }
@@ -78,16 +78,20 @@ class LocationManager {
   }
 
   Future<void> stopService() async {
-    if (locationManager != null) {
+    try {
       locationManager.pause();
       print('Stoped');
+    } catch (e) {
+      print('error...');
     }
   }
 
   Future<void> resumeService() async {
-    if (locationManager != null) {
+    try {
       locationManager.resume();
       print('Resuming...');
+    } catch (e) {
+      print('error...');
     }
   }
 }
